requirements.txt
streamlit==1.29.0
pandas==2.1.0
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
twilio==8.10.0
groq
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
schedule==1.2.1